add file uploading and other stuff

+++ nfweb.py
@@ -1,9 +1,11 @@
+import subprocess
 import pathlib
 import typing
+
 import os
 import signal
 
-from flask import Flask, request, render_template, redirect
+from flask import Flask, request, render_template, redirect, abort
 
 import nflib
 import config
@@ -25,23 +27,34 @@ def chunks(l, n):
 
 @app.route('/')
 def list_flows():
-    flows = list()
-    for nf in cfg.get('nextflows'):
-        flows.append(nf)
-    return render_template('list_flows.template', flows=flows)
+    return render_template('list_flows.template', flows=cfg.get('nextflows'))
 
-@app.route('/flow/<flow_name>/new')
+@app.route('/flow/<flow_name>/new', methods=['GET', 'POST'])
 def begin_run(flow_name: str):
-    flow_cfg = flows[flow_name]
-    flow_input_cfg = flow_cfg['input'][0] # <- hmm
-    return render_template('begin_run.template', flow=flow_cfg, incfg=flow_input_cfg)
+    if request.method == 'GET':
+        flow_cfg = flows[flow_name]
+        flow_input_cfg = flow_cfg['input']
+        return render_template('begin_run.template', flow=flow_cfg, incfg=flow_input_cfg)
+
+    elif request.method == 'POST':
+        flow_cfg = flows[flow_name]
+        flow_input_cfg = flow_cfg['input']
+        vs = []
+        for v in request.form.values():
+            vs.append(v)
+        cmd_line = flow_input_cfg['script']
+        input_str = flow_input_cfg['argf'].format(*vs)
+        print(cmd_line, input_str)
+
+        os.system("nohup bash {0} {1} &".format(cmd_line, input_str))
+        return redirect("/flow/{0}".format(flow_name))
 
 @app.route('/list_all')
 def list_all_runs():
     data = list()
     for flow in flows:
         nf_directory = pathlib.Path(flow['directory'])
-        table = nflib.parseHistoryFile(str(nf_directory / 'history'))
+        table = nflib.parseHistoryFile(nf_directory / 'history')
         table.reverse()
         datum = { 'table': table, 'flow_name': flow['name'] }
         data.append(datum)
@@ -50,9 +63,12 @@ def list_all_runs():
 @app.route('/flow/<flow_name>')
 def list_runs(flow_name : str):
     data = list()
-    flow_cfg = flows[flow_name]
+    try:
+        flow_cfg = flows[flow_name]
+    except:
+        abort(404)
     nf_directory = pathlib.Path(flow_cfg['directory'])
-    table = nflib.parseHistoryFile(str(nf_directory / 'history'))
+    table = nflib.parseHistoryFile(nf_directory / 'history')
     table.reverse()
     datum = { 'table': table, 'flow_name': flow_name }
     data.append(datum)
@@ -83,7 +99,9 @@ def run_details(flow_name : str, run_uuid: int):
         buttons['dagdot'] = True
 
     trace_filename = nf_directory / 'traces/{0}.trace'.format(run_uuid)
-    trace_nt = nflib.parseTraceFile(str(trace_filename))
+    if not trace_filename.is_file():
+        abort(404)
+    trace_nt = nflib.parseTraceFile(trace_filename)
     return render_template('run_details.template', uuid=run_uuid, flow_name=flow_name, entries=trace_nt, buttons=buttons)
 
 @app.route('/flow/<flow_name>/log/<run_uuid>')
@@ -92,9 +110,8 @@ def show_log(flow_name : str, run_uuid: int):
     content = None
     with open(str(log_filename)) as f:
         content = f.read()
-    
+
     return render_template('show_log.template', content=content, flow_name=flow_name, uuid=run_uuid)
-    pass
 
 @app.route('/flow/<flow_name>/report/<run_uuid>')
 def show_report(flow_name : str, run_uuid: int):
@@ -117,12 +134,14 @@ def show_dagdot(flow_name: str, run_uuid: int):
 @app.route('/flow/<flow_name>/stop/<run_uuid>')
 def kill_nextflow(flow_name : str, run_uuid: int):
     pid_filename = pathlib.Path(flows[flow_name]['directory']) / 'pids' / pathlib.Path("{0}.pid".format(run_uuid)).name
-    pid = None
-    with open(str(pid_filename)) as f:
-        pid = int(f.readline())
-    if pid:
-        try:
-            os.kill(pid, signal.SIGTERM)
-        except:
-            pass
+    if pid_filename.is_file():
+        pid = None
+        with open(str(pid_filename)) as f:
+            pid = int(f.readline())
+        if pid:
+            try:
+                print("killing pid {0}".format(pid))
+                os.kill(pid, signal.SIGTERM)
+            except:
+                pass
     return redirect("/flow/{0}/details/{1}".format(flow_name, run_uuid), code=302)
